Reference

This section describes the API in detail.

Main Struct

VSPGeom.VSPComponent — Type

VSPComponent()

Parameters defining the VSPComponent object.

Arguments

type::String: Type

name::String: Name

SurfNdx::Int: Surface index

GeomID::String: Geometry ID

MainSurfNdx::Int: Main surface index

SymCopyNdx::Int: Symmetry copy index

surface_node::DataFrame: Surface node DegenGeom

surface_face::DataFrame: Surface face DegenGeom

plate::DataFrame: Plate DegenGeom

stick_node::DataFrame: Stick node DegenGeom

stick_face::DataFrame: Stick face DegenGeom

point::DataFrame: Point DegenGeom

VSPGeom.readDegenGeom — Function

readDegenGeom(filename::String; verbose::Bool=false)

Read DegenGeom CSV file written out by OpenVSP to obtain geometry and components.

Arguments

filename::String: DegenGeom filename

verbose::Bool: Set to true to print status messages during file read operation

Returns

comp: Vector of VSPComponent objects

VSPGeom.degenGeomSize — Function

degenGeomSize(degenGeom::DataFrame)

Get size of the 2d mesh data structure that describes the degenerate geometry. This is different from the DataFrame size obtained using size(degenGeom) which is the number of variables and total number of points in the DataFrame.

Arguments

degenGeom::DataFrame: One of the degenGeom inside the VSPComponent object

Returns

nx::Int: Usually represents the number of cross-sections, referred to as Xsecs in OpenVSP

ny::Int: Usually represents the number of points in each cross-section in OpenVSP